+++ Tree_Controller/Components/Vegetation.cs
@@ -5,11 +5,25 @@
 namespace Tree_Controller
 {
     using Unity.Entities;
+    using Unity.Mathematics;
 
     /// <summary>
-    /// A component used to find out Vegetation quickly with a query.
+    /// A component used to find Vegetation prefabs quickly with a query.
     /// </summary>
     public struct Vegetation : IComponentData, IQueryTypeParameter
     {
+        /// <summary>
+        /// Records the original size from ObjectGeometryData Component.
+        /// </summary>
+        public float3 m_Size;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Vegetation"/> struct.
+        /// </summary>
+        /// <param name="size">Records the original size from ObjectGeometryData component.</param>
+        public Vegetation(float3 size)
+        {
+            m_Size = size;
+        }
     }
 }