+++ Tree_Controller/Systems/FindTreesAndBushesSystem.cs
@@ -122,15 +122,42 @@ namespace Tree_Controller.Systems
 
             foreach (UIGroupElement element in buffer)
             {
+                if (EntityManager.HasComponent<Vegetation>(element.m_Prefab))
+                {
+                    continue;
+                }
+
                 EntityManager.AddComponent<Vegetation>(element.m_Prefab);
+                if (EntityManager.TryGetComponent(element.m_Prefab, out ObjectGeometryData objectGeometryData))
+                {
+                    m_Log.Debug($"{nameof(FindTreesAndBushesSystem)}.{nameof(OnGameLoadingComplete)} objectGeometryData.m_size = {objectGeometryData.m_Size.x}:{objectGeometryData.m_Size.z}");
+                    Vegetation vegetation = new Vegetation(new Unity.Mathematics.float3(objectGeometryData.m_Size.x, 0, objectGeometryData.m_Size.z));
+                    EntityManager.SetComponentData(element.m_Prefab, vegetation);
+
+                    if (TreeControllerMod.Instance.Settings.LimitedTreeAnarchy)
+                    {
+                        objectGeometryData.m_Size.x = objectGeometryData.m_LegSize.x;
+                        objectGeometryData.m_Size.z = objectGeometryData.m_LegSize.z;
+                        EntityManager.SetComponentData(element.m_Prefab, objectGeometryData);
+                    }
+                }
+                else
+                {
+                    if (m_PrefabSystem.TryGetPrefab(element.m_Prefab, out PrefabBase prefabBase))
+                    {
+                        m_Log.Info($"{nameof(FindTreesAndBushesSystem)}.{nameof(OnGameLoadingComplete)} A vegetation prefab: {prefabBase.name} didn't have Object Geometry Data.'");
+                    }
+                    else
+                    {
+                        m_Log.Info($"{nameof(FindTreesAndBushesSystem)}.{nameof(OnGameLoadingComplete)} An unknown vegetation prefab didn't have Object Geometry Data.'");
+                    }
+                }
             }
 
             if (TreeControllerMod.Instance.Settings.FreeVegetation)
             {
                 m_ModifyVegetationPrefabsSystem.SetVegetationCostsToZero();
             }
-
-            m_ModifyVegetationPrefabsSystem.DecreaseObjectGeometryBounds();
 
             base.OnGameLoadingComplete(purpose, mode);
         }