+++ Tree_Controller/ChangeLog.md
@@ -3,4 +3,6 @@
 * Added Dead trees back to age selection with object tool. (Not Line Tool for now.)
 * Added Free Trees and Vegetation option. (Default Off)
 * Added option to add Stumps to age selection with object tool and tree changing tools. (Default Off)
-* Added option to constrain brush while adding trees and vegetation with object tool brush mode. (Default On).
+* Added option to constrain brush while adding trees and vegetation with object tool brush mode. (Default On)
+* Added option for Faster Max Brush Strength. 300% replaces 100%. (default off).
+* Added option for Limited Tree Anarchy, which makes conflict checks for trees only based apporimxately on their trunks. (Default off)